Add citus_job_{list,status} API
citus_job_list() is a simple wrapper UDF that simply shows the records
in pg_dist_background_job, and this contains all the rows in the
associated table.

citus_job_status(job_id) is a UDF that returns all the rows in the
former UDF, and an extra details column that shows various details on
the task.